bun_cc: `gnu_inline` means GNU's `inline` in Microsoft C too; four fixtures that assumed a platform
With `extern inline __attribute__((gnu_inline))` it is GNU's `inline` that was asked for, whatever the dialect: compiling for Windows the definition was taken for one of Microsoft's inline functions, and the definition proper that may follow it was a redefinition.

Tests: a complex fixture is written without `<complex.h>`, which Microsoft's library has another one of; a bit-field of 40 bits is a `long long`; the test that prints the name of an included file takes it after either separator.

+++ test/js/bun/ffi/bir/fixtures/structs-and-abi/structures-with-nothing-to-pass.c
@@ -6,7 +6,7 @@
 
 struct nothing {};
 struct only_padding { int : 8; };
-struct more_padding { long : 40; int : 0; };
+struct more_padding { long long : 40; int : 0; };
 struct of_nothing { struct nothing inside; } __attribute__((aligned(8)));
 struct nearly { struct only_padding padding; char one; };
